Shriram Pune - Sit Back, Relax. Your New View Awaits
Shriram Pune is the first child-centric residential project by Shriram Properties, located in Pune. It emphasizes the physical, mental, cultural, and frivolous development of kids from an early age. This township project covers massive wide hectares of land parcel with approximately wide hectares of open space and is well equipped with all the contemporary features and facilities that are a prerequisite to nurturing a child towards a brighter tomorrow. It has child-centric avenues within its building for the overall development of a child that includes- The Club, The Culture Corner, The Fit Hub, The Attention Lab & Worry-free Adults. 

Shriram Apartments In Pune offers 1, 2 & 3 BHK Luxury Homes with luxury finishes and fittings. All the residences in this housing enclave are provided with rooms that have vitrified tiles and balconies & toilets are provided with anti-skid ceramic tiles. The main arrival door has a teak frame with a flush door shutter with veneer & melamine elegance. All the homes are provided with windows with mosquito mesh. 

Shriram Flats in Pune enjoys an exceptional location advantage. It is situated in one of the most sought-after locations in Pune owing to its connectivity quotient. E-city has one of the largest electronic/IT industrial parks which is spread across wide acres. Also, it is in proximity to other IT hubs & enjoys outstanding connectivity Junction. This venture lies close to the proposed Airport that is set to provide a fillip to the real estate market in the area. 

The prime factor driving new projects in Pune is the presence of IT & ITes industries. Also, E-city has been placed among the top hot residential property locations. All the exceptional location advantages joined with the brand trust & the booming real estate market in the area make Shriram Property Pune stand apart from all the other residential projects in the locality.

For More Details :
Visit - https://shriram.developerprojects.in/pune/shriram-pune/